move queue
