+++ jobs/mongodb_migration/tests/test_deletion_migrations.py
@@ -9,7 +9,7 @@ from libcommon.constants import (
     QUEUE_MONGOENGINE_ALIAS,
 )
 from libcommon.dtos import Status
-from libcommon.queue import JobDocument
+from libcommon.new_queue.jobs import JobDocument
 from libcommon.resources import MongoResource
 from libcommon.utils import get_datetime
 from mongoengine.connection import get_db
