+++ libs/libcommon/src/libcommon/orchestrator.py
@@ -18,9 +18,9 @@ from libcommon.constants import (
     DIFFICULTY_BONUS_BY_FAILED_RUNS,
 )
 from libcommon.dtos import JobInfo, JobResult, Priority
+from libcommon.new_queue.jobs import Queue
 from libcommon.processing_graph import ProcessingGraph, ProcessingStep, ProcessingStepDoesNotExist, processing_graph
 from libcommon.prometheus import StepProfiler
-from libcommon.queue import Queue
 from libcommon.simple_cache import (
     CachedArtifactNotFoundError,
     delete_dataset_responses,
