+++ services/admin/src/admin/routes/pending_jobs.py
@@ -6,8 +6,8 @@ from typing import Optional
 
 from libapi.exceptions import ApiError, UnexpectedApiError
 from libapi.utils import Endpoint, get_json_api_error_response, get_json_ok_response
+from libcommon.new_queue.jobs import Queue
 from libcommon.processing_graph import processing_graph
-from libcommon.queue import Queue
 from starlette.requests import Request
 from starlette.responses import Response
