+++ services/worker/tests/job_runners/config/test_parquet_and_info.py
@@ -31,7 +31,7 @@ from libcommon.dtos import JobInfo, JobParams, Priority
 from libcommon.exceptions import (
     DatasetManualDownloadError,
 )
-from libcommon.queue import Queue
+from libcommon.new_queue.jobs import Queue
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
